Rewards

Documentation for Rewards.

Rewards.jl is required internally only. It provides the correct reward shaping to be used by the MPC unit in the MBRL.jl Package. Since the planning unit has no access to the actual environment, it is neccessary to emit deterministic rewards from the model.

Installation

In the julia REPL, run

using Pkg
Pkg.add(url="https://github.com/SvenDuve/Rewards.jl")

Bring package into scope with ( not required ):

using Rewards

Function Reference

Rewards.Reward

Rewards.Reward

Rewards.Reward

Rewards.Reward

Rewards.Reward

Rewards.Reward

Rewards.Reward

Rewards.Reward

Rewards.Reward

Rewards.shaping

Rewards.shaping

Rewards.shaping

Rewards.shaping

Rewards.Reward — Type

Reward

Reward struct for the environments. It contains the environment, the previous shaping, the game_over flag and the awake flag.

Rewards.Reward — Method

(rew::Reward)(Environment::Acrobot, state, action, next_state)

Reward functor for the Acrobot environment.

Rewards.Reward — Method

Reward(Environment::Acrobot)

Constructor for the Reward struct for the Acrobot environment.

Rewards.Reward — Method

(rew::Reward)(Environment::LunarLanderContinuous, state, action, next_state)

Reward functor for the LunarLanderContinuous environment.

Rewards.Reward — Method

Reward(Environment::LunarLanderContinuous; prev_shaping = 0.f0, game_over = false, awake = true)

Constructor for the Reward struct for the LunarLanderContinuous environment.

Rewards.Reward — Method

(rew::Reward)(Environment::LunarLanderDiscrete, state, action, next_state)

Reward functor for the LunarLanderDiscrete environment.

Rewards.Reward — Method

Reward(Environment::LunarLanderDiscrete; prev_shaping = 0.f0, game_over = false, awake = true)

Constructor for the Reward struct for the LunarLanderDiscrete environment.

Rewards.Reward — Method

(rew::Reward)(Environment::Pendulum, state, action, next_state)

Reward functor for the Pendulum environment.

Rewards.Reward — Method

Reward(Environment::Pendulum)

Constructor for the Reward struct for the Pendulum environment.

Rewards.shaping — Method

shaping(Environment::Acrobot, s)

Shaping function for the Acrobot environment.

Rewards.shaping — Method

shaping(Environment::LunarLanderContinuous, s)

Shaping function for the LunarLanderContinuous environment.

Rewards.shaping — Method

shaping(Environment::LunarLanderDiscrete, s)

Shaping function for the LunarLanderDiscrete environment.

Rewards.shaping — Method

shaping(Environment::Pendulum, s)

Shaping function for the Pendulum environment.